bugs fixed.

+++ database/src/main/java/com/shukalovich/database/dao/ProductDao.java
@@ -1,37 +1,34 @@
 package com.shukalovich.database.dao;
 
 import com.shukalovich.database.DummyDatabase;
-import com.shukalovich.database.entity.Brand;
 import com.shukalovich.database.entity.Product;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class ProductDao implements Dao<Integer, Product>{
+public final class ProductDao implements Dao<Long, Product> {
     private static final ProductDao INSTANCE = new ProductDao();
     private final DummyDatabase db = DummyDatabase.getInstance();
-
     public static ProductDao getInstance() {
         return INSTANCE;
     }
 
     @Override
-    public Product delete(Integer id) {
+    public Product delete(Long id) {
         return Optional.ofNullable(db.getProducts().remove(id))
                 .orElseThrow(RuntimeException::new);
     }
 
     @Override
     public List<Product> findAll() {
-        return new ArrayList<> (db.getProducts().values());
+        return new ArrayList<>(db.getProducts().values());
     }
 
     @Override
-    public Optional<Product> findById(Integer id) {
+    public Optional<Product> findById(Long id) {
         return Optional.ofNullable(db.getProducts().get(id));
     }
